Add ability for Orka to take app and script as parameters
Instead of using config.ini to specify the location of the apps and
the monkeyrunner script, it now has support for directly specifying
it as command line parameters when the Orka process is run. This will
be useful for different processes calling Orka, as it no longer relies
on a config file which could be changed arbitarily from concurrent
requests.