Add ability to call droidmate to interact with app
Orka can now use Droidmate to automate the interactions with the apk
file to test. This is done by modifying simulationMaster to run the
droidmate JAR file that has been added to the dependencies folder. A new
command line parameter has been added to Orka to specify Orka to use
Droidmate. Also:
- Add extra environmental variables to paths.sh
- Add a comment to the monkey_recorder.py