Update inject.py to handle DroidMate instrumented apks
Methods injectMethodPrologue now checks for the end of the file and returns when the end is reached. Also the DroidMate instrumented apks no longer have the .prologue when the smali files are generated which means the function will carry on till the end of the file. To stop this, the function has been changed such that it now checks for the keyword 'droidmate'. injectMethodBody also updated to check for end of file too. Also:
- Update method names to Monkeyrunner and DroidMate-2 in main.py
- Slight update to comments in simulationMaster.sh for clarity