Fix stack error issue in Orka
Issue dealing with abstract method calls in Orka. Abstract methods do
not have a method enter or exit statements, as they are abstract,
but Orka may think they need one if their implementation is inside the
same smali file, and the implementation methods have API calls.
I fix this by checking if the invoked method uses the same parameter
structure as a method of interest (that is a method with the same
name and that we know contains an API call).

Also fixed an issue in the hardwareAnalyser. Many of the batterystats
for the process do not seem to have the cpu= part anymore. I have
now changed the regex to grab the number after the : symbol. If
this also cannot be found then the it will print out a message instead
of crashing.

Also fixed broken pipe messages in instrument.sh by adding a 2>/dev/null
to the command. This will ignore the SIGPIPE error messages.