Ignore bad execution traces
There seems to be a problem with Orka when it comes to apps
with more than one activity. The execution trace sometimes
will have extra exit statements that are either out of place
or do not have a corresponding enter. Some methods also
do not seem to have a corresponding exit. For now, I make a
check for the exit ones that appear out of place, and remove them
(and everything after) from the stack. Also clarified some comments
so that it's easier to understand what all the different stacks do.

In inject.py, we now add exit logs for throw statements.

Updated instrument.sh to work with apps that decompiles into
multiple smali folders. I made the assumption that it would only
ever be in one smali folder, but this is not the case, and so we need
to perform the injection into all the relevant decompiled folders.
Seems that the injection into smali_classes2 sometimes breaks
functionality of the app.

Also updated .gitignore