Update APIAnalyser to handle mismatched stacks
Sometimes the routine stack and the line stack may have different
sizes, so the call to addAPICall will raise an error. Looking at
the logcat files produced, it seems this may be caused when an
an API call happens directly after an invoke statement. Normally,
after an invoke statement there will a method enter statement, which
will append onto the routine stack. I work around this for now, by
passing the routine name to the function, so that it can append the
correct routineName first. Also:
- Made some of the error messages a bit more verbose
- it's also possible that the routineStack is empty when an
  erroneous exit statement occurs. In these situations, I just mark
  the expectedName as an empty string