Add extra command line parameters
Extra cmd parameters have been added to specify the AVD and port number
to use. netStatsMonitor and simulationMaster are then updated to specify
the emulator serial for all the adb commands if an emulator is being used.
simulationMaster in particular now uses the port number that is passed in as
a argument to the script. monkey_player_orka also updated to connect
to the specific emulator.