Fix instrumentation and injection issues
If instrument.sh fails due to the fact that it cannot find the package directory
within the smali folders then it will exit with exit code 2. helpers.py is then
updated such that if the exit code of the process ran was 2 it will return with
the phrase "Exit code 2". This allows processes that wanted to run that cmd
to handle specific exit codes. In particular, main.py handles failures of
exit code 2 by retrying the cmd but with different package directory
structures - it removes the first directory each time. This is in case
the package directory is not the full package name, such as with
relay app.

In inject.py, we fix a bug when getting parameters from the method
signature. Longs and doubles take up two parameters and as such
the next parameter along should skip a p register. During the
mapping from p registers to v registers, the code will account for
this.

Fix a bug when calling move operation with a register bigger than
v15. If that is the case, we change it to move/from16.

Also if the parameter is an array of objects, then when moving the
parameter back, we also use move-object/from16