Add additional error handling and limit smali folder check
Account for KeyErrors when removing items from routine stack.
When looking for the smali folder to inject into, initially limit
the directory to the top 2 levels. This is because most times the
files are stored in here, and the full package name may not exist
or the files in the full package directory will not have anything
injected, resulting in an empty logcat file.