Fix bug where smali code uses v registers for parameters
If Orka decides to inject in a method but deems it unnecessary to
perform any remapping of the parameters, then there could be a bug
where the smali code uses the v registers to point to the parameters.
Because Orka will now use an additional local register to hold the
API names, any code that used to reference the first parameter with
the v register will now point to the API name instead, which can
cause issues. For now, we inject an additional line to move the
first parameter back to the new local register, but this only accounts
for the first parameter. In the future, we need need to move all the
affected parameter registers back one, and then after the line has
executed, move them back.